Fix param types in LightSaml\Model\Assertion\Conditions class
The `setNotBefore` and `setNotOnOrAfter` methods actually accept `int|string|\DateTime`, and do not accept null, since the argument is passed directly to `Helper::getTimestampFromValue()`.

This fixes an "Expected int, got DateTime" warning in PhpStorm when passing a `DateTime` instance to the `setNotOnOrAfter` method.

Also corrected the return type of `getNotBeforeString()`.